pyproject.toml
[tool.poetry]
name = "stonks"
version = "0.0.1"
description = "A flask-based web application that allows users to manage a personalized stock watchlist"
authors = ["orhantugrul <orhantugrulsahin@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.3"
flask-sqlalchemy = "^3.1.1"
flask-wtf = "^1.2.1"
finnhub-python = "^2.4.20"
flask-login = "^0.6.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
